requirements.txt
absl-py==1.4.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
batchgenerators==0.21
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dominate==2.8.0
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.13.1
fonttools==4.33.3
fsspec==2023.12.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h5py==3.6.0
huggingface-hub==0.19.4
idna==3.3
imageio==2.21.1
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
linecache2==1.0.0
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
MedPy==0.4.0
mistune==0.8.4
mmengine==0.7.3
monai==1.0.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pip==21.2.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pynng==0.7.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
rawpy==0.18.1
regex==2023.6.3
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==68.1.2
shapely==2.0.1
SimpleITK==2.1.1.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.8.8
timm==0.6.13
tinycss2==1.1.1
tomli==2.0.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
ultralytics==8.0.159
unittest2==1.1.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.6
wheel==0.37.1
widgetsnbextension==3.6.0
yacs==0.1.8
yapf==0.33.0
zipp==3.16.2
